+++ src/server/tcp_server.py
@@ -5,7 +5,7 @@ import os
 from protocol.config_manager import ConfigManager
 from protocol.protocol_factory import ProtocolFactory
 
-from src.chatservice.api import (
+from src.services.api import (
     signup, login, delete_user, get_chats, get_all_users, update_view_limit,
     save_settings, start_chat, get_user_message_limit, delete_chats, delete_messages, get_messages, send_chat_message, get_users_to_display
 )
